requirements.txt
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
channels==4.2.0
chardet==5.2.0
Django==5.1.7
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
docopt==0.6.2
et_xmlfile==2.0.0
greenlet==3.1.1
num2words==0.5.14
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
reportlab==4.3.1
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
